refactor bounds checking and add loop push check
